+++ app/settings/extensions/page.tsx
@@ -3,8 +3,8 @@ import { SidebarPage } from '@/components/layout/SidebarPage';
 import axios from 'axios';
 import { getCookie } from 'cookies-next';
 import { useSearchParams } from 'next/navigation';
-import { useAgent } from '@/components/idiot/interactive/hooks/useAgent';
-import { useCompany } from '@/components/idiot/useUser';
+import { useAgent } from '@/components/interactive/useAgent';
+import { useCompany } from '@/components/interactive/useUser';
 import { useEffect, useState } from 'react';
 import OAuth2Login from 'react-simple-oauth2-login';
 import { providers as oAuth2Providers } from '@/components/auth/OAuth';
