+++ components/interactive/useUser.ts
@@ -1,5 +1,5 @@
-import { chainMutations, createGraphQLClient } from '@/components/idiot/interactive/hooks/lib';
-import '@/components/idiot/zod2gql/zod2gql';
+import { chainMutations, createGraphQLClient } from '@/components/interactive/lib';
+import '@/components/interactive/zod2gql';
 import axios from 'axios';
 import { getCookie } from 'cookies-next';
 import useSWR, { SWRResponse } from 'swr';
