+++ components/layout/agent-selector.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { useCompany } from '@/components/idiot/useUser';
+import { useCompany } from '@/components/interactive/useUser';
 import {
   DropdownMenu,
   DropdownMenuContent,
@@ -14,7 +14,7 @@ import { setCookie } from 'cookies-next';
 import { Check, ChevronsUpDown, Plus } from 'lucide-react';
 import { useRouter } from 'next/navigation';
 import { FaRobot } from 'react-icons/fa';
-import { Agent, useAgent, useAgents } from '@/components/idiot/interactive/hooks/useAgent';
+import { Agent, useAgent, useAgents } from '@/components/interactive/useAgent';
 
 export function AgentSelector() {
   const { isMobile } = useSidebar('left');
